refactored code base for cleaner dependency management
