+++ leapp/export_manager.py
@@ -31,7 +31,7 @@ from leapp.leapp_graph.datatypes import (
     is_traced_type,
     is_tracable_tensor_type,
 )
-from leapp.leapp_graph.datatypes.global_patching import warn_if_script_functions_in_scope
+from leapp.leapp_graph.datatypes.patching import warn_if_script_functions_in_scope
 from leapp.utils.tensor_description import TensorSemantics
 from leapp.utils.tensor_description import (verify_data_exact_match,
                                              flatten_io_structure,
