+++ leapp/leapp_graph/custom_operator_registry/warp_operator/bundle.py
@@ -3,9 +3,48 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-"""Pack a Warp APIC ``.wrp`` graph plus its ``_modules/`` directory into one blob.
+"""Self-contained Warp APIC bundle serialization and FX graph embedding.
 
-Archive layout matches ``experiment_warp/onnx_embedded_wrp/wrp_bundle.py`` (WRPB v1).
+Purpose
+-------
+During LEAPP trace, each Warp segment's APIC graph is saved to a **temporary**
+``.wrp`` file (plus a sibling ``*_modules/`` tree) by ``TracedTensorNode``.
+That layout is fine on the build machine but is not portable inside an exported
+``.pt2`` or ``.onnx`` artifact.
+
+This module closes that gap in two steps:
+
+1. **WRPB archive** (``pack_bundle``, ``build_archive``) — read ``segment.wrp``
+   and its ``segment_modules/`` directory from disk and pack them into a single
+   byte blob (magic ``WRPB``, version 1). The format matches the experiment
+   prototype in ``experiment_warp/onnx_embedded_wrp/wrp_bundle.py``.
+
+2. **FX embed pass** (``embed_warp_bundles_in_graph``) — before
+   ``torch.export`` or ONNX dynamo export, replace each ``leapp::warp_runner``
+   node's zero-length bundle placeholder with a ``get_attr`` to a registered
+   ``uint8`` buffer holding the WRPB bytes. The exported graph is then
+   self-contained: no external ``.wrp`` paths at inference time.
+
+Pipeline position
+-----------------
+::
+
+    trace (WarpOp)  →  save .wrp to temp dir (traced_node)
+                    →  pack WRPB + wire into FX graph (this module, pre_compile)
+                    →  export .pt2 / .onnx (warp_operator lowering)
+                    →  ONNX-only: patch WrpRunner string attrs (onnx_export_backend)
+
+``iter_warp_segments_from_graph`` is a small helper shared with save/ONNX passes
+to list ``WarpSegment`` objects in ``warp_runner`` node order.
+
+Why a separate module
+---------------------
+``warp_operator`` owns op registration, fake/eager kernels, and ONNX
+lowering. This file owns **on-disk → in-graph bytes** wiring only. They could
+be merged into one export-support module, but keeping WRPB serialization here
+avoids coupling a binary archive format to ``torch.library`` registration.
+``datatypes.warp.warp_segment`` holds segment *state*; this module handles
+segment *payload* embedding at export time.
 """
 
 from __future__ import annotations
@@ -76,9 +115,9 @@ def _is_warp_runner_node(node: fx.Node, warp_op) -> bool:
 
 def iter_warp_segments_from_graph(graph: fx.Graph) -> list:
     """Return warp segment objects in ``warp_runner`` node order."""
-    from leapp.leapp_graph.custom_operator_registry import warp_custom_op
+    from .schema import get_op
 
-    warp_op = warp_custom_op.get_op().default
+    warp_op = get_op().default
     segments = []
     for node in graph.nodes:
         if not _is_warp_runner_node(node, warp_op):
@@ -97,10 +136,10 @@ def embed_warp_bundles_in_graph(graph_module: fx.GraphModule) -> int:
     ``node.meta['leapp_warp_segment']`` in graph order. Called from the Warp
     ``pre_compile`` hook before ``onnx-dynamo`` or ``exported-program`` export.
     """
-    from leapp.leapp_graph.custom_operator_registry import warp_custom_op
+    from .schema import get_op
     from leapp.utils.logging import _get_logger
 
-    warp_op = warp_custom_op.get_op().default
+    warp_op = get_op().default
     warp_nodes = [
         node for node in graph_module.graph.nodes if _is_warp_runner_node(node, warp_op)
     ]
