+++ leapp/leapp_graph/datatypes/torch/__init__.py
@@ -1,0 +1,8 @@
+#
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+
+from .traced_tensor import TracedTensor
+
+__all__ = ["TracedTensor"]
